# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dot Test Sprint 2",
        "version": "0.1"
    },
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Authorize user",
                "description": "Authorizes user by its email and password",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Authentication successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "authorization token",
                                            "type": "string",
                                            "example": "fSPJ2AR0TU0dLB6aiYgtSGHkPnFTfBdh4ltISiSo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/v1/search/cities": {
            "get": {
                "tags": [
                    "City"
                ],
                "summary": "Get Data City",
                "description": "Returns data city",
                "operationId": "GetCities",
                "parameters": [
                    {
                        "name": "origins",
                        "in": "path",
                        "description": "to switchable data source",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success get data city"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/search/provinces": {
            "get": {
                "tags": [
                    "Province"
                ],
                "summary": "Get Data Province",
                "description": "Returns data Province",
                "operationId": "GetProvinces",
                "parameters": [
                    {
                        "name": "origins",
                        "in": "path",
                        "description": "to switchable data source",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success get data province"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "properties": {
                    "email": {
                        "description": "User Email",
                        "type": "string",
                        "format": "email",
                        "nullable": false
                    },
                    "password": {
                        "description": "User Password",
                        "type": "string",
                        "example": "password",
                        "nullable": false
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Bearer Token Authentication",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}